I made some modifications to RSofiaFacade.cc specifically with how the matrix was being indexed.
Also, the predict method compiles now and I didn't have to do much.

I think all I did was put "std::" in front of the string and put "RSofiaFacade::" in front of predict method in the .cc file

I doubt it is doing what it should be doing accurately, because i think some things are screwy with the indexing but we can talk about that later.